added blueprint functionality
